fixing np float, and saving raw data

+++ gunpowder/nodes/precomputed_source.py
@@ -101,7 +101,7 @@ class PrecomputedSource(BatchProvider):
         spec.dtype = vol.dtype
 
         spec.interpolatable = spec.dtype in [
-            np.float,
+            float,
             np.float32,
             np.float64,
             np.float128,
